+++ projects/slap_manipulation/scripts/test_detic_on_obs.py
@@ -1,81 +1,43 @@
-import glob
-import os
-from typing import List
-
-import click
-import h5py
 import numpy as np
+import rospy
 from matplotlib import pyplot as plt
+from PIL import Image
 
-import home_robot.utils.data_tools.image as image
 from home_robot.core.interfaces import Observations
 from home_robot.perception.detection.detic.detic_perception import DeticPerception
+from home_robot.utils.data_tools.loader import DatasetBase
+from home_robot.utils.data_tools.writer import DataWriter
+from home_robot_hw.env.stretch_manipulation_env import StretchManipulationEnv
 
 MY_CATEGORIES = ["cup", "bottle", "drawer", "basket", "bowl", "computer", "mug"]
-TASK_TO_OBJECT_MAP = {
-    "open-top-drawer": ["drawer handle"],
-    "close-top-drawer": ["drawer handle"],
-}
 
 
-def sandwich(obj_list: List[str]):
-    return ["other"] + obj_list + ["other"]
+class SimpleDataset(DatasetBase):
+    def get_datum(self, trial, idx):
+        # return two dicts: data, config
+        pass
 
 
-@click.command()
-@click.option("--data-dir", type=str, default="~/data/dataset.h5")
-@click.option("--template", type=str, default="*/*.h5")
-@click.option(
-    "--mode", type=click.Choice(["read", "write"], case_sensitive=True), default="read"
-)
-def main(data_dir, template, mode):
+if __name__ == "__main__":
+    rospy.init_node("test_detic")
+    robot = StretchManipulationEnv(init_cameras=True)
     segmentation = DeticPerception(
         vocabulary="custom",
         custom_vocabulary=",".join(MY_CATEGORIES),
         sem_gpu_id=0,
     )
-    depth_factor = 10000
-    files = glob.glob(os.path.join(data_dir, template))
-    for file in files:
-        # get object category to look for given task
-        if mode == "read":
-            h5file = h5py.File(file, "r")
-        else:
-            h5file = h5py.File(file, "a")
-        if mode == "write":
-            task_name = h5file[list(h5file.keys())[0]]["task_name"][()].decode("utf-8")
-            object_for_task = TASK_TO_OBJECT_MAP[task_name]
-        # segmentation.reset_vocab(sandwich(object_for_task))
-        for g_name in h5file.keys():
-            if mode == "read":
-                plt.imshow(h5file[g_name]["head_semantic_frame"][()])
-                plt.show()
-            else:
-                if "head_semantic_frame" not in h5file[g_name].keys():
-                    rgb = image.img_from_bytes(h5file[g_name]["head_rgb/0"][()])
-                    depth = (
-                        image.img_from_bytes(h5file[g_name]["head_depth/0"][()])
-                        / depth_factor
-                    )
-                    detic_obs = Observations(
-                        rgb=rgb,
-                        depth=None,
-                        xyz=h5file[g_name]["head_xyz"][()][0],
-                        gps=np.zeros(2),  # TODO Replace
-                        compass=np.zeros(1),  # TODO Replace
-                        task_observations={},
-                    )
-                    result = segmentation.predict(detic_obs)
-                    h5file[g_name]["head_semantic_frame"] = result.task_observations[
-                        "semantic_frame"
-                    ]
-                    h5file[g_name]["head_semantic_mask"] = result.semantic
-                    plt.imshow(result.task_observations["semantic_frame"])
-                    plt.show()
-                else:
-                    print(f"Already processed {g_name} in {file}")
-        h5file.close()
-
-
-if __name__ == "__main__":
-    main()
+    # image = Image.open("./desk.jpg")
+    # image_np = np.array(image)
+    obs = robot.get_observation()
+    breakpoint()
+    detic_obs = Observations(
+        rgb=obs["rgb"],
+        depth=obs["depth"],
+        xyz=obs["xyz"],
+        gps=np.zeros(2),  # TODO Replace
+        compass=np.zeros(1),  # TODO Replace
+        task_observations={},
+    )
+    result = segmentation.predict(detic_obs)
+    plt.imshow(result.task_observations["semantic_frame"])
+    plt.show()
